# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NoorAl Quran Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: List all bookmarks for the authenticated user
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Bookmark an ayah
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Notes ──────────────────────────────────────────────────────────────────
...
  /hard-ayahs:
    get:
      operationId: getHardAyahs
      tags: [hard-ayahs]
      summary: Get hard ayahs due for review
      responses:
        "200":
          description: List of hard ayahs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HardAyah"
    post:
      operationId: flagHardAyah
      tags: [hard-ayahs]
      summary: Flag an ayah as hard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HardAyahInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardAyah"
  /hard-ayahs/{id}/review:
    post:
      operationId: reviewHardAyah
      tags: [hard-ayahs]
      summary: Submit a review result for a hard ayah (spaced repetition)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HardAyahReview"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardAyah"
  /hard-ayahs/{id}:
    delete:
      operationId: removeHardAyah
      tags: [hard-ayahs]
      summary: Remove an ayah from the hard list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Memorization ───────────────────────────────────────────────────────────
...
  /memorization/ayahs:
    post:
      operationId: markAyahMemorized
      tags: [memorization]
      summary: Mark an ayah as memorized or not
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorizationEntry"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorizationProgress"
...
  /goals:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get current reading goal
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingGoal"
    put:
      operationId: saveGoal
      tags: [goals]
      summary: Create or update reading goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadingGoalInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingGoal"
  /goals/progress:
    get:
      operationId: getGoalProgressHistory
      tags: [goals]
      summary: Get last 30 days of goal progress
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalProgress"
    post:
      operationId: recordGoalProgress
      tags: [goals]
      summary: Record today's reading progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalProgressInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalProgress"

  # ── Streaks ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Bookmark:
      type: object
      required: [id, userId, surahId, ayahNumber, surahNameEn, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        surahNameAr:
          type: string
        ayahText:
          type: string
        createdAt:
          type: string
    BookmarkInput:
      type: object
      required: [surahId, ayahNumber, surahNameEn]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        surahNameAr:
          type: string
        ayahText:
          type: string

    # ── Notes ─────────────────────────────────────────────────────────────────
...
    HardAyah:
      type: object
      required: [id, userId, surahId, ayahNumber, nextReviewDate, reviewCount]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        nextReviewDate:
          type: string
        reviewCount:
          type: integer
        lastResult:
          type: ["string", "null"]
    HardAyahInput:
      type: object
      required: [surahId, ayahNumber]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
    HardAyahReview:
      type: object
      required: [result]
      properties:
        result:
          type: string
          enum: [correct, needs_review]

    # ── Memorization ──────────────────────────────────────────────────────────
    MemorizationProgress:
      type: object
      required: [id, userId, surahId, ayahNumber, memorized]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        memorized:
          type: boolean
        memorizedAt:
          type: ["string", "null"]
    MemorizationEntry:
      type: object
      required: [surahId, ayahNumber, memorized]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        memorized:
          type: boolean
...
    ReadingGoal:
      type: object
      required: [id, userId, goalType, targetAmount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        goalType:
          type: string
          enum: [ayahs, pages, juz]
        targetAmount:
          type: integer
        createdAt:
          type: string
    ReadingGoalInput:
      type: object
      required: [goalType, targetAmount]
      properties:
        goalType:
          type: string
          enum: [ayahs, pages, juz]
        targetAmount:
          type: integer
    GoalProgress:
      type: object
      required: [id, userId, date, amountRead, goalMet]
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        amountRead:
          type: integer
        goalMet:
          type: boolean
    GoalProgressInput:
      type: object
      required: [amountRead]
      properties:
        amountRead:
          type: integer

    # ── Streaks ───────────────────────────────────────────────────────────────
...